Add new location of Oncosis
OncoCis is a user friendly tool for annotating cancer mutations in cis-regulatory regions of DNA. This tutorial will provide a step by step guide to using this web service. It will also give a detailed description of the expected output and the various options and features that will potentially be useful for identifying possible causal mutations in cis-regulatory regions.